feat: rebuild site using Gatsby
Pretty much none of the projects this site relied upon (Bower,
Metalsmith, etc.) are maintained anymore.

This commit rebuilds the site using Gatsby. It also removes the
dependency on Zurb Foundation in favor of hand-rolled CSS; Foundation is
a huge framework and I was only ever using a tiny slice of it, mostly
for things that are now much easier to accomplish like grid layouts.